cleanup command

+++ flox_bootstrap/project.py
@@ -1,40 +1,45 @@
 import os
 import re
 import shutil
+from os.path import isdir
 from pathlib import Path
 from shutil import copy2
 
 import stringcase
-from jinja2 import Environment, FileSystemLoader
-
 from floxcore import CONFIG_DIRS
-from floxcore.console import warning
+from floxcore.console import warning, error_box
 from floxcore.context import Flox
 from floxcore.remotes import universal_copy, generate_cache_hash
+from jinja2 import Environment, FileSystemLoader
 
 
 def _reload_cache(flox: Flox, cache_dir: str):
-    shutil.rmtree(cache_dir)
+    if isdir(cache_dir):
+        shutil.rmtree(cache_dir)
 
     for repository in flox.settings.bootstrap.repositories:
         universal_copy(flox, cache_dir, repository)
 
 
-def enable(flox: Flox, bootstrap: tuple, no_cache: bool, **kwargs):
+def enable(flox: Flox, features: tuple, no_cache: bool, **kwargs):
     """Bootstraps project with given templates"""
     cache_dir = CONFIG_DIRS.get_in("user", "templates-cache")
 
-    if no_cache:
+    if not len(flox.settings.bootstrap.repositories):
+        error_box("Repositories not configured. Run: flox configure --plugin=bootstrap")
+        return
+
+    if no_cache or not isdir(cache_dir):
         _reload_cache(flox, cache_dir)
 
     existing_paths = []
-    for template_name in bootstrap:
+    for template_name in features:
         for repository in flox.settings.bootstrap.repositories:
             template_path = os.path.join(cache_dir, generate_cache_hash(repository), template_name)
             if os.path.isdir(template_path):
                 existing_paths.append(template_path)
 
-    non_existing = set(bootstrap) - set([Path(p).parts[-1] for p in existing_paths])
+    non_existing = set(features) - set([Path(p).parts[-1] for p in existing_paths])
     for name in non_existing:
         warning(f'Bootstrap "{name}" does not exist')
